added table variables and index update
I added all the variables into the adresses table and a bunch more into the user table. There is still some like the profile picture that will be added when we decide how to store it. I also added into the profile page the starts to work based off of the users rating, so if the rating is 3 then 3 stars will appear