added seperate password form for editting user info
Now there is a second button in the user info column do update password. Added the method into the controller and created the needed routes. However I am having trouble with the validation so it wont update the password yet. The ui is slightly weird as the columns arent equal distance apart but that will get changed soon when we use tailwind so whateva